+++ src/utility/tdbcompare.c
@@ -312,10 +312,10 @@ main (int argc, char *argv [])
 
       if (strcmp (tag [DB_A], tagdefs [TAG_DURATION].tag) == 0) {
         if (val [DB_A] != NULL && val [dblocidx] != NULL) {
-          int32_t     dura, durb;
+          int64_t     dura, durb;
 
-          dura = atol (val [DB_A]);
-          durb = atol (val [dblocidx]);
+          dura = atoll (val [DB_A]);
+          durb = atoll (val [dblocidx]);
           if (abs (dura - durb) > 100) {
             fprintf (stderr, "    tdbcomp:k: tag %s mismatch /%s/%s/ %s\n", tag [DB_A], val [DB_A], val [dblocidx], fn);
             grc = 1;
